refactor: Update navigation routes to use Get.offAllNamed instead of Get.toNamed
This commit updates the navigation routes in the loginController.dart and userData.dart files to use Get.offAllNamed instead of Get.toNamed. This change ensures that the user is completely navigated to the specified route, removing any previous routes from the navigation stack. By using Get.offAllNamed, we provide a smoother and more intuitive navigation experience for the users.