refactor: Ignore test files in .gitignore
This commit updates the .gitignore file to ignore the newly added test files, `lib/test.dart` and `lib/test2.dart`. By excluding these files from version control, we ensure that they are not included in the repository's history. This refactor improves the organization of the repository and prevents unnecessary clutter.